requirements.txt
# Core dependencies
alpaca-py>=0.28.0
robin-stocks>=3.4.0
webull-python-sdk-core>=0.1.18
webull-python-sdk-trade>=0.1.14
APScheduler>=3.10.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytz>=2024.1

# Email providers (optional)
sendgrid>=6.10.0
boto3>=1.34.0

# Persistence (optional)
firebase-admin>=6.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
responses>=0.24.0
freezegun>=1.2.0
